fix: better handle endoflife.date api "not found"
The endoflife.date api returns a 404 for multiple condidions, including

- A product is requested which doesn't exist in the api
- An existing product is requested, but the version isn't registered

This challenges an existing assumption that you could throw any
real version number at the api and get a result. In reality. old
versions for real products may still return a 404.

This fix allows processing to continue for a 404, but prints out
a message indicating that there was a problem.